feat:project cards are clickable
