+++ synergysphere/src/app/page.tsx
@@ -1,189 +1,103 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import { Project, CreateProjectData, TeamMember } from "@/types/project";
-import { ProjectCard } from "@/components/project-card";
-import { NewProjectModal } from "@/components/new-project-modal";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Search, Users, FolderOpen } from "lucide-react";
-
-const mockProjects: Project[] = [
-  {
-    id: "1",
-    name: "E-commerce Platform",
-    description: "Building a modern e-commerce platform with React and Node.js. Features include user authentication, product catalog, shopping cart, and payment integration.",
-    progress: 75,
-    teamMembers: [
-      { id: "1", name: "John Doe", email: "john@example.com", initials: "JD" },
-      { id: "2", name: "Jane Smith", email: "jane@example.com", initials: "JS" },
-      { id: "3", name: "Mike Johnson", email: "mike@example.com", initials: "MJ" },
-    ],
-    createdAt: new Date("2024-01-15"),
-    updatedAt: new Date("2024-01-20"),
-  },
-  {
-    id: "2",
-    name: "Mobile App Redesign",
-    description: "Complete redesign of our mobile application to improve user experience and add new features.",
-    progress: 45,
-    teamMembers: [
-      { id: "4", name: "Sarah Wilson", email: "sarah@example.com", initials: "SW" },
-      { id: "5", name: "David Brown", email: "david@example.com", initials: "DB" },
-    ],
-    createdAt: new Date("2024-01-10"),
-    updatedAt: new Date("2024-01-18"),
-  },
-  {
-    id: "3",
-    name: "Data Analytics Dashboard",
-    description: "Creating an interactive dashboard for business intelligence and data visualization.",
-    progress: 90,
-    teamMembers: [
-      { id: "6", name: "Emily Davis", email: "emily@example.com", initials: "ED" },
-      { id: "7", name: "Chris Lee", email: "chris@example.com", initials: "CL" },
-      { id: "8", name: "Lisa Garcia", email: "lisa@example.com", initials: "LG" },
-      { id: "9", name: "Tom Wilson", email: "tom@example.com", initials: "TW" },
-    ],
-    createdAt: new Date("2024-01-05"),
-    updatedAt: new Date("2024-01-19"),
-  },
-];
+import Image from "next/image";
 
 export default function Home() {
-  const [projects, setProjects] = useState<Project[]>(mockProjects);
-  const [searchTerm, setSearchTerm] = useState("");
-  const [filteredProjects, setFilteredProjects] = useState<Project[]>(mockProjects);
-
-  useEffect(() => {
-    const filtered = projects.filter(project =>
-      project.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      project.description.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-    setFilteredProjects(filtered);
-  }, [projects, searchTerm]);
-
-  const handleCreateProject = (data: CreateProjectData) => {
-    const teamMembers: TeamMember[] = data.teamMemberEmails.map((email, index) => ({
-      id: `new-${index}`,
-      name: email.split('@')[0],
-      email,
-      initials: email.split('@')[0].substring(0, 2).toUpperCase(),
-    }));
-
-    const newProject: Project = {
-      id: Date.now().toString(),
-      name: data.name,
-      description: data.description,
-      progress: 0,
-      teamMembers,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    };
-
-    setProjects(prev => [newProject, ...prev]);
-  };
-
-  const handleProjectClick = (project: Project) => {
-    console.log("Opening project:", project.name);
-  };
-
   return (
-    <div className="min-h-screen bg-background">
-      <header className="border-b bg-card">
-        <div className="container mx-auto px-4 py-6">
-          <div className="flex items-center justify-between">
-            <div className="flex items-center gap-3">
-              <div className="h-8 w-8 bg-primary rounded-lg flex items-center justify-center">
-                <Users className="h-5 w-5 text-primary-foreground" />
-              </div>
-              <div>
-                <h1 className="text-2xl font-bold">SynergySphere</h1>
-                <p className="text-sm text-muted-foreground">Team Collaboration Platform</p>
-              </div>
-            </div>
-            <NewProjectModal onCreateProject={handleCreateProject} />
-          </div>
-        </div>
-      </header>
+    <div className="font-sans grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
+      <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
+        <Image
+          className="dark:invert"
+          src="/next.svg"
+          alt="Next.js logo"
+          width={180}
+          height={38}
+          priority
+        />
+        <ol className="font-mono list-inside list-decimal text-sm/6 text-center sm:text-left">
+          <li className="mb-2 tracking-[-.01em]">
+            Get started with the Odoo X NMIT Hackathon{" "}
+            <code className="bg-black/[.05] dark:bg-white/[.06] font-mono font-semibold px-1 py-0.5 rounded">
+              src/app/page.tsx
+            </code>
+            .
+          </li>
+          <li className="tracking-[-.01em]">
+            Save and see your changes instantly.
+          </li>
+        </ol>
 
-      <main className="container mx-auto px-4 py-8">
-        <div className="mb-8">
-          <h2 className="text-3xl font-bold mb-2">Welcome back!</h2>
-          <p className="text-muted-foreground">
-            Manage your projects and collaborate with your team in one place.
-          </p>
-        </div>
-
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-8">
-          <div className="bg-card p-6 rounded-lg border">
-            <div className="flex items-center gap-3">
-              <FolderOpen className="h-8 w-8 text-primary" />
-              <div>
-                <p className="text-2xl font-bold">{projects.length}</p>
-                <p className="text-sm text-muted-foreground">Total Projects</p>
-              </div>
-            </div>
-          </div>
-          <div className="bg-card p-6 rounded-lg border">
-            <div className="flex items-center gap-3">
-              <Users className="h-8 w-8 text-primary" />
-              <div>
-                <p className="text-2xl font-bold">
-                  {projects.reduce((acc, project) => acc + project.teamMembers.length, 0)}
-                </p>
-                <p className="text-sm text-muted-foreground">Team Members</p>
-              </div>
-            </div>
-          </div>
-          <div className="bg-card p-6 rounded-lg border">
-            <div className="flex items-center gap-3">
-              <div className="h-8 w-8 bg-green-500 rounded-full flex items-center justify-center">
-                <span className="text-white text-sm font-bold">✓</span>
-              </div>
-              <div>
-                <p className="text-2xl font-bold">
-                  {projects.filter(p => p.progress === 100).length}
-                </p>
-                <p className="text-sm text-muted-foreground">Completed</p>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="flex items-center gap-4 mb-6">
-          <div className="relative flex-1 max-w-md">
-            <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-            <Input
-              placeholder="Search projects..."
-              value={searchTerm}
-              onChange={(e) => setSearchTerm(e.target.value)}
-              className="pl-10"
+        <div className="flex gap-4 items-center flex-col sm:flex-row">
+          <a
+            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:w-auto"
+            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Image
+              className="dark:invert"
+              src="/vercel.svg"
+              alt="Vercel logomark"
+              width={20}
+              height={20}
             />
-          </div>
+            Deploy now
+          </a>
+          <a
+            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 w-full sm:w-auto md:w-[158px]"
+            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Read our docs
+          </a>
         </div>
-
-        {filteredProjects.length === 0 ? (
-          <div className="text-center py-12">
-            <FolderOpen className="h-12 w-12 text-muted-foreground mx-auto mb-4" />
-            <h3 className="text-lg font-semibold mb-2">No projects found</h3>
-            <p className="text-muted-foreground mb-4">
-              {searchTerm ? "Try adjusting your search terms." : "Get started by creating your first project."}
-            </p>
-            {!searchTerm && <NewProjectModal onCreateProject={handleCreateProject} />}
-          </div>
-        ) : (
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-6">
-            {filteredProjects.map((project) => (
-              <ProjectCard
-                key={project.id}
-                project={project}
-                onClick={() => handleProjectClick(project)}
-              />
-            ))}
-          </div>
-        )}
       </main>
+      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/file.svg"
+            alt="File icon"
+            width={16}
+            height={16}
+          />
+          Learn
+        </a>
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/window.svg"
+            alt="Window icon"
+            width={16}
+            height={16}
+          />
+          Examples
+        </a>
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/globe.svg"
+            alt="Globe icon"
+            width={16}
+            height={16}
+          />
+          Go to nextjs.org →
+        </a>
+      </footer>
     </div>
   );
 }
